Reverted 63960aa and 60540c2.
These changes collectively caused outfits to have shorter cooldown
periods when you didn't use them as long. But this change was made
with the intention of also adding 6a4bf10, which was rejected upstream.
Without that, all this does is create an advantage to doing several
short bursts of activating the outfit, or even having some sort of
keyboard macro do it for you. I don't think there's any possible
reason to want to implement this kind of thing on purpose.