long distance cargo missions
————Summary—————
Added a mission chain that culminates in a new repeatable “long
distance” shipping missions. These missions require going from empire
space to other faction space, and require 3 to 6 jumps. In return, they
pay a bit more per jump and give a little more rep per mission. As with
the other empire shipping missions, the deadlines are easy to make.
This is the first campaign I’ve scripted, and it was mostly
cannibalized from the previous empire shipping line.

———Extended Details————
The first mission has a 30% chance of appearing once the player
complete the empire recruitment. The player flies bureaucrats to 5
factions, and then takes Lieutenant Czesc (of the first empire shipping
mission fame) back to Gamma Polaris. The factions are, in order:
Soromid, Dvaered, Za’lek, Frontier, and Sirius. After the first
mission, each mission has a 75% chance of appearing in a bar in Empire
space, except the last mission, which has a 100% chance of appearing.
Afterwards, long-distance shipping routes appear in the mission
computer, requiring at least 3 jumps and ending in non-Empire space.

The idea is to encourage pilots to explore the universe a bit,
although, to make it easier, I picked stations at the border of a given
territory. The long-distance routes are more lucrative in terms of
payment (compared to regular ES missions) and add a little more danger
because they go into non-Empire controlled space.